+++ tests/sakura_parser_test.cpp
@@ -22,9 +22,10 @@
 
 #include "sakura_parser_test.h"
 
-#include <libKitsunemimiSakuraLang/items/sakura_items.h>
 #include <parsing/sakura_parsing.h>
-#include <sakura_garden.h>
+
+#include <libKitsunemimiSakuraLang/items/sakura_items.h>
+#include <libKitsunemimiSakuraLang/sakura_garden.h>
 
 #include <test_strings/test_strings.h>
